Implement up cloud login command
up cloud login recieves either a user or token and fetches a session
token for it before storing in the up config file. The session token is
good for 30 days before a user will be required to login again. If this
is the first time a user is logging in, the user or token profile will
be set as the default to be used by other commands. A future command
will allow for adjusting the default profile.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>