Strip leading v off of UXP version when pulling chart
Strips the leading v off of any UXP versions supplied when pulling the
chart so that it matches how Helm caches it. An alternative approach
would be to parse into structured semantic version prior to passing to
Helm, but we prefer to defer that logic to Helm rather than parsing into
semver twice. If additional modification were made by Helm in the
future, this functionality could break, but we would prefer to hit the
issue rather than potentially pulling and installing incorrect UXP versions.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>