Cache latest chart with proper version on install and upgrade
Currently, when a version is not specified on uxp install or upgrade, we
pull the latest stable chart and cache it without version information in
the dedicated uxp chart cache. This means that if a version that was
installed as the latest was subsequently installed explicitly (i.e. by
providing the version), we would unnecessarily pull the chart again when
it was already in the cache.

This updates the pull logic to save charts pulled as latest with their
appropriate version so that they can be identified in future installs.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>