Migrate from github.com/dgrijalva/jwt-go to github.com/golang-jwt/jwt
Updates the library used for interacting with JWTs as the previously
used package contains a vulnerability and is no longer maintained. Note
that the vulnerability is not particularly impactful in this tool
specifically, but we prefer to go ahead and move to a package that is
actively maintained.

See CVE-2020-26160.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>