Update helm install manager to be more flexible
Updates the helm install manager to be less coupled to installing UXP.
This is accomplished by allowing a caller to specify an alternate chart
rather an assuming Crossplane is the alternate. It also adds new options
for providing user / pass authentication and ability to specify chart is
hosted in an OCI registry. Instead of switching on an unstable flag,
caller is now required to pass the repoURL, meaning that using the
stable or unstable chart URL is now decided outside of the helm install
manager. This decouples the unstable behavior from helm's --devel flag,
which we always set given that we do not want to restrict the versions a
user can pull by tag, but rather by channel (url).

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>