Add up enterprise upgrade command
Adds the up enterprise upgrade command, which behaves similarly to up
enterprise install, but it requires that the enterprise chart is already
installed before attempting upgrade. Users are also prompted for license
ID and token in this command, though if the chart is already present in
the cache then they could be supplied as empty.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>