Refactor internal config package to use upbound instead of cloud
Updates the internal config structure and methods to use upbound instead
of cloud to refer to configuration that is related to interacting with
Upbound. This is a breaking change despite being an internal package,
but it will only require users logging in again, at which point their
existing config file will be overwritten with the revised structure.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>