Update install and upgrade to both apply license and image pull secrets
up enterprise update will now create or update the image pull secret and
the license secret using the credentials passed by the user. This is
necessary because a new access key is required for any new version of
enterprise, and the image pull secret may need new credentials if a
different license id is being used.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>