Add initial workspace implemenation
Adds an initial workspace implementation that contains a cache of nodes
and validators. Workspace parsing can either be done at the workspace or
file level. In the future we will want to support incremental parsing
using text change notifications from the LSP client. We will also want
to support reparsing nodes based on changes to nodes they depend on.
This can be accomplished by constructing a graph of all nodes (objects)
in a workspace, then traversing the edges to a node when it is updated.

Validators are currently constructured from the package cache, but
assume the on-disk representation is CRDs, rather than xpkg tarballs.
This will need to be adjusted. There is also currently only support for
v1 CRDs, but we will need to support both v1beta1 CRDs and all versions
of XRDs for validation schemas.

A workspace is thread-safe, but currently use coarse-grained locking,
which can likely be optimized in the future.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>