Parse and validate embedded resources in a Composition
Adds recursive parsing of embedded resources within a Composition. As
part of this change documents are unmarshalled to unstructured during
parsing, rather than waiting until validation. This ensures that we do
attempt to parse embedded resources when the parent Composition is not
well formed and also reduces the number of path lookups we have to
perform for acquiring GVK.

Embedded resources are declared as dependants of their parent, but exist
as standalone package nodes in the workspace. This allows for future
validation granularity, such as only validating a package node and the
resources that depend on it, which may be embedded resources, or in the
future, resources that depend on the schema defined by the resource.

Diagnostic generation from validation errors has also been separated
from the broader validation logic. A follow up will implement
diagnostics on failed parsing.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>