Do not unmarshal in validation
Updates to remove the marshal / unmarshal dance in validation because we
perform it ahead of time in parsing, which makes the object available to
source GVK and other metadata from in validation.

Note that exposing metav1.Object and the ast.Node means that these
structures in a PackageNode could be modified outside of parsing. We may
want to consider placing some additional safeguards around the methods
exposed by the Node interface in the future to make it such that
underlying structures are immutable outside of construction. On the
other hand, being able to update these structures in memory could serve
useful when we move to processing events such as keystrokes.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>