Add support for --output on xpkg build and deprecate --name
Adds suppport for --output on xpkg build and deprecates support for
--name. If neither flag is given, the package directory is used as
output location and the name is generated (metadata.name plus hash),
which matches existing behavior. Command errors if both --name and
--output are supplied. If --output is supplied it is used verbatim after
cleaning path (i.e. there is no generated file name and the .xpkg
extension is not automatically added).

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>