Only pass domain to docker-credential-up
up requires that commands provide a scheme in the UP_DOMAIN, but doing
so causes a failure to match to registry domain in docker-credential-up,
meaning that a separate UP_DOMAIN value must be used for up and
docker-credential-up. This corrects that behavior to parse the hostname
from UP_DOMAIN in docker-credential-up and only pass it along as the
domain for matching.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>