Move controlplanes and upbound under alpha but maintain root
Moves controlplanes and upbound commands under the alpha distinction,
but also maintains hidden location at root in the tree. All
documentation and references will use the `up alpha` variants, but the
hidden commands serve to reserve a future location for these commands,
and, in this case, preserve backwards compatability with commands that
already existed.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>