Restructure alpha command gating
Restructures alpha level command gating to dynamically hide commands
from help based on the branch they are accessed on. This allows for
tagging commands such that they can be accessible from multiple
branches, but are not shown in the help at maturity levels that differ
from their own.

Also updates to hide the usage help on hidden commands, instead
encouraging folks to use the variant at the specified maturity level. In
the future we will introduce alpha commands and not allow them to be
invoked on the main tree, but for now we want to keep existing commands
functioning properly even though they are not discoverable.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>