spaces: Fix the account name check in `up space attach`
Previously, the check in `up space attach` was taking the current account name
only from the `--account` parameter, rather than using the existing machinery
that takes it from the active profile in the config file when necessary. Use the
account name from the upbound context, which should alwyas be populated
appropriately.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>