ctp: Allow version and upgrade channel to be specified in create
Add flags to specify the desired Crossplane version and upgrade channel in the
`up ctp create` command. Return an error if a version is specified with a
release channel other than `None`, since the server will accept a spec with a
version specified but ignore it in favor of the release channel's current
version.

For now, we don't have a programmatic way to fetch the valid release channels so
we hard-code the options. Similarly, we don't have an API to fetch supported
versions so we can't list them for the user in the CLI.

Fixes #541

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>