ctx: Make cursor placement logic more robust
We programmatically manipulate the cursor in two cases:

1. When constructing the list of items, to place the cursor below the "back"
   item in the list if it exists.
2. When the cursor comes across an unselectable item, to "skip" it.

We now need to also handle unselectable items in case (1), and case (2) didn't
correctly handle lists with unselectable items at the beginning or end.

Consolidate this handling by introducing some helper functions that handle the
cursor movement and correctly handle the corner cases. We handle unselectable
items at the beginning or end of lists by "bouncing" back in the direction the
user was moving until we hit a selectable item; in the corner case where all
items are unselectable we'll end up with an unselectable item selected, which is
probably the best we can do given the constraints of the list library.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>