ctx: Update the list asynchronously and add a spinner
Update the item list asynchronously to the main update loop of the TUI and
trigger an update once we have our items. This lets us run a spinner while the
items are being updated, which makes the UI feel more responsive when it's
taking a few seconds to fetch the items.

As part of this change, switch from generating our own title to using the list
widget's built-in title bar. This doesn't change how the TUI looks at all, but
lets us leverage the built-in spinner capability of the list widget.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>