ctx: Show unselectable items for unreachable Spaces
Avoid the surprise of seeing different Spaces in the console vs. `up ctx` by
listing unreachable items for Spaces that are unreachable or for which we failed
to find the ingress. Sort these items to the end of the list so that all the
selectable items are at the top.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>