ctx: Correctly handle errors when the user selects an item
Previously, if we hit an error in the `onEnter` function when a user selected an
item, we would get stuck completely: the spinner would keep spinning forever. If
we hit an error when listing items for a new screen, the error would be
displayed but the user wouldn't be able to navigate.

Fix both issues by correctly handling the errors. In both cases, we re-enable
the keymap and return the existing model, so the user can continue to navigate.

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>