ctx: Add space name to kubeconfig extension
Previously, we parsed the ingress URL to determine the name of the Space when
deriving initial state. This was correct for cloud spaces, but not for connected
spaces, since they use a URL starting with `proxy.`.

Add the name of the space to our kubeconfig extension and use it as the space
name when deriving state if it's present. If it's not present, fall back to the
old behavior (in case anyone has a pre-existing kubeconfig with the old
extension in it).

Fixes #557

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>