ctx: Ignore empty path components in non-interactive mode
Allow for trailing slashes (and also, as a side-effect, duplicate slashes) in
non-interactive mode by ignoring any empty path components when we parse the
path. This is especially helpful since we print a trailing slash on the path
when a user selects a context other than a control plane (space or group).

Fixes #581

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>