project: Correctly ignore non-KRM YAML files in build
`up project build` is intended to ignore any YAML files that don't contain an
XRD or composition. We do this by parsing the YAML to get its `apiVersion` and
`kind` and collect only the relevant files into an in-memory filesystem that we
hand off to the package parser (which will choke if it encounters an unexpected
type).

Previously, we were parsing the YAMLs as `unstructured.Unstructured`; this
throws an error if the `kind` field isn't found, defeating the purpose of our
pre-parsing. This wasn't caught by the tests due to the bug fixed in the
previous commit.

Parse into a `TypeMeta` instead, since it holds all the fields we care about
anyway and doesn't care if `kind` is missing. While we're here, add validation
in the tests to ensure we have the right number of resources in the
package (i.e., didn't ignore ones we should include).

Signed-off-by: Adam Wolfe Gordon <awg@upbound.io>